feat: rename json_lines to jsonl for cli
closes: https://github.com/influxdata/influxdb/issues/25873
